+++ test/Marketplace/updateDirectListing.ts
@@ -3,15 +3,15 @@ import { expect, use } from "chai";
 import { solidity } from "ethereum-waffle";
 
 // Contract Types
-import { MockERC1155 } from "../../../typechain/MockERC1155";
-import { Marketplace, ListingParametersStruct, ListingStruct } from "../../../typechain/Marketplace";
+import { MockERC1155 } from "../../typechain/MockERC1155";
+import { Marketplace, ListingParametersStruct, ListingStruct } from "../../typechain/Marketplace";
 
 // Types
 import { BigNumber } from "ethers";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
 
 // Test utils
-import { getContracts, Contracts } from "../../../utils/tests/getContracts";
+import { getContracts, Contracts } from "../../utils/tests/getContracts";
 
 use(solidity);
 
@@ -51,18 +51,6 @@ describe("Edit listing: direct listing", function () {
   const approveMarketToTransferTokens = async (toApprove: boolean) =>
     await mockNft.connect(lister).setApprovalForAll(marketv2.address, toApprove);
 
-  const timeTravelToListingWindow = async (listingId: BigNumber) => {
-    // Time travel
-    const listingStart: string = (await marketv2.listings(listingId)).startTime.toString();
-    await ethers.provider.send("evm_mine", [parseInt(listingStart)]);
-  };
-
-  const timeTravelToAfterListingWindow = async (listingId: BigNumber) => {
-    // Time travel
-    const listingEnd: string = (await marketv2.listings(listingId)).endTime.toString();
-    await ethers.provider.send("evm_mine", [parseInt(listingEnd)]);
-  };
-
   before(async () => {
     // Get signers
     const signers: SignerWithAddress[] = await ethers.getSigners();
@@ -98,7 +86,7 @@ describe("Edit listing: direct listing", function () {
       reservePricePerToken: ethers.utils.parseEther("0.1"),
       buyoutPricePerToken: ethers.utils.parseEther("0.2"),
 
-      listingType: ListingType.Auction,
+      listingType: ListingType.Direct,
     };
 
     // Setup: `lister` lists nft for sale in a direct listing.
@@ -106,27 +94,6 @@ describe("Edit listing: direct listing", function () {
   });
 
   describe("Revert cases", function () {
-    it("Should revert if the auction has already started", async () => {
-      await timeTravelToListingWindow(listingId);
-
-      const newStartTime: BigNumber = (listingParams.startTime as BigNumber).add(100);
-      const newSecondsUntilEndTime: BigNumber = BigNumber.from(5000);
-
-      await expect(
-        marketv2
-          .connect(lister)
-          .updateListing(
-            listingId,
-            listingParams.quantityToList,
-            listingParams.reservePricePerToken,
-            listingParams.buyoutPricePerToken,
-            listingParams.currencyToAccept,
-            newStartTime,
-            newSecondsUntilEndTime,
-          ),
-      ).to.be.revertedWith("Marketplace: auction already started.");
-    });
-
     it("Should revert if lister edits quantity to an amount they don't own or have approved for transfer", async () => {
       const invalidNewQuantity: BigNumber = nftTokenSupply.add(1);
 
@@ -145,59 +112,10 @@ describe("Edit listing: direct listing", function () {
           ),
       ).to.be.revertedWith("Marketplace: insufficient token balance or approval.");
     });
-
-    it("Should revert if caller is not auction creator", async () => {
-      const newStartTime: BigNumber = (listingParams.startTime as BigNumber).add(100);
-      const newSecondsUntilEndTime: BigNumber = BigNumber.from(5000);
-
-      await expect(
-        marketv2
-          .connect(protocolAdmin)
-          .updateListing(
-            listingId,
-            listingParams.quantityToList,
-            listingParams.reservePricePerToken,
-            listingParams.buyoutPricePerToken,
-            listingParams.currencyToAccept,
-            newStartTime,
-            newSecondsUntilEndTime,
-          ),
-      ).to.be.revertedWith("Marketplace: caller is not listing creator.");
-    });
   });
 
   describe("Events", function () {
-    it("Should emit AuctionClosed if new listing quantity is 0", async () => {
-      const newQuantity: BigNumber = BigNumber.from(0);
-      const newStartTime: BigNumber = (listingParams.startTime as BigNumber).add(100);
-      const newSecondsUntilEndTime: BigNumber = BigNumber.from(5000);
-
-      await expect(
-        marketv2
-          .connect(lister)
-          .updateListing(
-            listingId,
-            newQuantity,
-            listingParams.reservePricePerToken,
-            listingParams.buyoutPricePerToken,
-            listingParams.currencyToAccept,
-            newStartTime,
-            newSecondsUntilEndTime,
-          ),
-      )
-        .to.emit(marketv2, "AuctionClosed")
-        .withArgs(
-          ...Object.values({
-            listingId: listingId,
-            closer: lister.address,
-            cancelled: true,
-            auctionCreator: lister.address,
-            winningBidder: ethers.constants.AddressZero,
-          }),
-        );
-    });
-
-    it("Should emit ListingUpdate with new listing info, if new listing quantity is not 0", async () => {
+    it("Should emit ListingUpdate with new listing info", async () => {
       const newStartTime: BigNumber = (listingParams.startTime as BigNumber).add(100);
       const newSecondsUntilEndTime: BigNumber = BigNumber.from(5000);
 
@@ -220,13 +138,10 @@ describe("Edit listing: direct listing", function () {
   });
 
   describe("Balances", function () {
-    it("Should update token balance of Marketplace and lister on editing listing quantity", async () => {
-      const amountToRemove: BigNumber = BigNumber.from(1);
-      const newListingQuantity = (listingParams.quantityToList as BigNumber).sub(amountToRemove);
+    it("Should not affect token balance on editing listing quantity", async () => {
+      const newListingQuantity = (listingParams.quantityToList as BigNumber).sub(1);
 
-      const listerBalBefore: BigNumber = await mockNft.balanceOf(lister.address, nftTokenId);
-      const marketBalBefore: BigNumber = await mockNft.balanceOf(marketv2.address, nftTokenId);
-
+      const balBefore: BigNumber = await mockNft.balanceOf(lister.address, nftTokenId);
       await marketv2
         .connect(lister)
         .updateListing(
@@ -238,41 +153,14 @@ describe("Edit listing: direct listing", function () {
           listingParams.startTime,
           listingParams.secondsUntilEndTime,
         );
-      const listerBalAfter: BigNumber = await mockNft.balanceOf(lister.address, nftTokenId);
-      const marketBalAfter: BigNumber = await mockNft.balanceOf(marketv2.address, nftTokenId);
+      const balAfter: BigNumber = await mockNft.balanceOf(lister.address, nftTokenId);
 
-      expect(listerBalAfter).to.equal(listerBalBefore.add(amountToRemove));
-      expect(marketBalAfter).to.equal(marketBalBefore.sub(amountToRemove));
+      expect(balAfter).to.equal(balBefore);
     });
   });
 
   describe("Contract state", function () {
-    it("Should cancel auction i.e. delete listing info if new quantity is 0", async () => {
-      const newQuantity: BigNumber = BigNumber.from(0);
-      const newStartTime: BigNumber = (listingParams.startTime as BigNumber).add(100);
-      const newSecondsUntilEndTime: BigNumber = BigNumber.from(5000);
-
-      await marketv2
-        .connect(lister)
-        .updateListing(
-          listingId,
-          newQuantity,
-          listingParams.reservePricePerToken,
-          listingParams.buyoutPricePerToken,
-          listingParams.currencyToAccept,
-          newStartTime,
-          newSecondsUntilEndTime,
-        );
-
-      const listing: ListingStruct = await marketv2.listings(listingId);
-
-      expect(listing.tokenOwner).to.equal(ethers.constants.AddressZero);
-    });
-
     it("Should store the edited listing state", async () => {
-      const amountToRemove: BigNumber = BigNumber.from(1);
-      const newListingQuantity = (listingParams.quantityToList as BigNumber).sub(amountToRemove);
-
       const newStartTime: BigNumber = (listingParams.startTime as BigNumber).add(100);
       const newSecondsUntilEndTime: BigNumber = BigNumber.from(6000);
 
@@ -280,7 +168,7 @@ describe("Edit listing: direct listing", function () {
         .connect(lister)
         .updateListing(
           listingId,
-          newListingQuantity,
+          listingParams.quantityToList,
           listingParams.reservePricePerToken,
           listingParams.buyoutPricePerToken,
           listingParams.currencyToAccept,
@@ -296,12 +184,12 @@ describe("Edit listing: direct listing", function () {
       expect(listing.tokenId).to.equal(listingParams.tokenId);
       expect(listing.startTime).to.equal(newStartTime);
       expect(listing.endTime).to.equal(newStartTime.add(newSecondsUntilEndTime));
-      expect(listing.quantity).to.equal(newListingQuantity);
+      expect(listing.quantity).to.equal(listingParams.quantityToList);
       expect(listing.currency).to.equal(listingParams.currencyToAccept);
       expect(listing.reservePricePerToken).to.equal(listingParams.reservePricePerToken);
       expect(listing.buyoutPricePerToken).to.equal(listingParams.buyoutPricePerToken);
       expect(listing.tokenType).to.equal(TokenType.ERC1155);
-      expect(listing.listingType).to.equal(ListingType.Auction);
+      expect(listing.listingType).to.equal(ListingType.Direct);
     });
   });
 });
